+++ lib/userstate.h
@@ -31,19 +31,19 @@ struct userstate_services {
 	                              void (*tick)   (WIDGET *, int dx, int dy),
 	                              void (*release)(WIDGET *, int dx, int dy));
 	void    (*grab)    (WIDGET *, void (*tick)   (WIDGET *, int dx, int dy));
-	long    (*get)               (void);
+	int    (*get)               (void);
 	void    (*handle)            (EVENT *, int);
 	WIDGET *(*get_mfocus)        (void);
 	void    (*set_active_window) (WINDOW *, int force);
 	WIDGET *(*get_selected)      (void);
-	long    (*get_mx)            (void);
-	long    (*get_my)            (void);
-	long    (*get_mb)            (void);
-	void    (*set_pos)           (long x, long y);
-	long    (*get_keystate)      (long keycode);
-	char    (*get_ascii)         (long keycode);
-	void    (*set_max_mx)        (long max_mx);
-	void    (*set_max_my)        (long max_my);
+	int    (*get_mx)            (void);
+	int    (*get_my)            (void);
+	int    (*get_mb)            (void);
+	void    (*set_pos)           (int x, int y);
+	int    (*get_keystate)      (int keycode);
+	char    (*get_ascii)         (int keycode);
+	void    (*set_max_mx)        (int max_mx);
+	void    (*set_max_my)        (int max_my);
 	void    (*release_app)       (int app_id);
 	void    (*release_widget)    (WIDGET *);
 };
